+++ services/autoscaling/tests/unit/test_utils_buffer_machines.py
@@ -20,7 +20,7 @@ from simcore_service_autoscaling.modules.cluster_scaling._provider_computational
 from simcore_service_autoscaling.modules.cluster_scaling._provider_dynamic import (
     DynamicAutoscalingProvider,
 )
-from simcore_service_autoscaling.utils.buffer_machines_pool_core import (
+from simcore_service_autoscaling.utils.buffer_machines import (
     dump_pre_pulled_images_as_tags,
     get_activated_buffer_ec2_tags,
     get_deactivated_buffer_ec2_tags,
